Send correct refresh rate on binding
Wayland unit of measurement for refresh rate is mHz, multiply by 1000
to send the correct value when the client is bound.

Also use qreal for the refreshRate because the API deals with Hz.

+++ src/compositor/compositor_api/qwaylandoutput.h
@@ -102,7 +102,7 @@ public:
     struct Mode
     {
         QSize size;
-        int refreshRate;
+        qreal refreshRate;
     };
 
     QWaylandOutput();
